+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples.2/FixedWidthColumns.jsx
@@ -12,12 +12,12 @@ const FeaturedIcon = () => {
     isFeatured ? <Button variant="utility" text="Featured button" icon={<IconFeatured />} onClick={onButtonClick} /> : <Button variant="utility" text="Featured off button" icon={<IconFeaturedOff />} onClick={onButtonClick} />
   );
 };
+
 const iconResultsNormal = <IconMultipleResultsNormal a11yLabel="Results normal" height="1.5em" width="1.5em" />;
 const iconResultsNotNormal = <IconMultipleResultsNotNormal a11yLabel="Results not normal" height="1.5em" width="1.5em" />;
 const iconResultsCritical = <IconMultipleResultsCritical a11yLabel="Results critical" height="1.5em" width="1.5em" />;
 
-// Source data for tests
-
+// Source rows data for tests
 const rows = [
   {
     id: 'row_1',
@@ -65,25 +65,25 @@ const cols = [
   {
     id: 'Column-0',
     displayName: 'Icon',
-    width: '60px',
+    width: '5em',
     align: alignTypes.CENTER,
   },
   {
     id: 'Column-1',
     displayName: 'Service name',
-    flexGrow: true,
+    width: '20em',
   },
   {
     id: 'Column-2',
     displayName: 'Featured',
-    width: '60px',
+    width: '5em',
     align: alignTypes.CENTER,
   },
 ];
 
 const FixedWidthColumns = () => (
   <CompactInteractiveList
-    id="compact-interactive-list-fixed-width-columns2"
+    id="interactive-compact-columns-width"
     ariaLabel="Compact Interactive List"
     rows={rows}
     columns={cols}